+++ lib/LibeventReactor.php
@@ -2,9 +2,7 @@
 
 namespace Amp;
 
-class LibeventReactor implements SignalReactor {
-    use GeneratorResolver;
-
+class LibeventReactor extends CoroutineResolver implements SignalReactor {
     private $base;
     private $watchers = [];
     private $immediates = [];
@@ -87,7 +85,7 @@ class LibeventReactor implements SignalReactor {
                 );
                 $result = $callback($this, $watcherId);
                 if ($result instanceof \Generator) {
-                    $this->resolveGenerator($result)->when($this->onCallbackResolution);
+                    $this->coroutine($result)->when($this->onCallbackResolution);
                 }
             } catch (\Exception $e) {
                 $this->handleRunError($e);
@@ -224,7 +222,7 @@ class LibeventReactor implements SignalReactor {
                 $watcherId = $watcher->id;
                 $result = $callback($this, $watcherId);
                 if ($result instanceof \Generator) {
-                    $this->resolveGenerator($result)->when($this->onCallbackResolution);
+                    $this->coroutine($result)->when($this->onCallbackResolution);
                 }
                 $this->cancel($watcherId);
             } catch (\Exception $e) {
@@ -289,7 +287,7 @@ class LibeventReactor implements SignalReactor {
             try {
                 $result = $callback($this, $watcherId);
                 if ($result instanceof \Generator) {
-                    $this->resolveGenerator($result)->when($this->onCallbackResolution);
+                    $this->coroutine($result)->when($this->onCallbackResolution);
                 }
 
                 // If the watcher cancelled itself this will no longer be set
@@ -363,7 +361,7 @@ class LibeventReactor implements SignalReactor {
             try {
                 $result = $callback($this, $watcherId, $stream);
                 if ($result instanceof \Generator) {
-                    $this->resolveGenerator($result)->when($this->onCallbackResolution);
+                    $this->coroutine($result)->when($this->onCallbackResolution);
                 }
             } catch (\Exception $e) {
                 $this->handleRunError($e);
@@ -411,7 +409,7 @@ class LibeventReactor implements SignalReactor {
             try {
                 $result = $callback($this, $watcherId, $signo);
                 if ($result instanceof \Generator) {
-                    $this->resolveGenerator($result)->when($this->onCallbackResolution);
+                    $this->coroutine($result)->when($this->onCallbackResolution);
                 }
             } catch (\Exception $e) {
                 $this->handleRunError($e);