+++ test/ReactorTest.php
@@ -2,6 +2,9 @@
 
 namespace Amp\Test;
 
+use Amp\Success;
+use Amp\Failure;
+
 abstract class ReactorTest extends \PHPUnit_Framework_TestCase {
     abstract protected function getReactor();
 
@@ -272,14 +275,12 @@ abstract class ReactorTest extends \PHPUnit_Framework_TestCase {
     }
 
     public function testOnStartGeneratorResolvesAutomatically() {
-        $reactor = $this->getReactor();
         $test = '';
-        $gen = function($reactor) use (&$test) {
-            yield;
+        $this->getReactor()->run(function($reactor) use (&$test) {
+            yield "pause" => 1;
             $test = "Thus Spake Zarathustra";
             $reactor->once(function() use ($reactor) { $reactor->stop(); }, 50);
-        };
-        $reactor->run($gen);
+        });
         $this->assertSame("Thus Spake Zarathustra", $test);
     }
 
@@ -287,7 +288,7 @@ abstract class ReactorTest extends \PHPUnit_Framework_TestCase {
         $reactor = $this->getReactor();
         $test = '';
         $gen = function($reactor) use (&$test) {
-            yield;
+            yield "pause" => 1;
             $test = "The abyss will gaze back into you";
             $reactor->once(function() use ($reactor) { $reactor->stop(); }, 50);
         };
@@ -300,7 +301,7 @@ abstract class ReactorTest extends \PHPUnit_Framework_TestCase {
         $reactor = $this->getReactor();
         $test = '';
         $gen = function($reactor) use (&$test) {
-            yield;
+            yield "pause" => 1;
             $test = "There are no facts, only interpretations.";
             $reactor->once(function() use ($reactor) { $reactor->stop(); }, 50);
         };
@@ -314,7 +315,7 @@ abstract class ReactorTest extends \PHPUnit_Framework_TestCase {
         $test = '';
         $gen = function($reactor, $watcherId) use (&$test) {
             $reactor->cancel($watcherId);
-            yield;
+            yield "pause" => 1;
             $test = "Art is the supreme task";
             $reactor->stop();
         };
@@ -376,4 +377,328 @@ abstract class ReactorTest extends \PHPUnit_Framework_TestCase {
         $this->assertSame(3, $var1);
         $this->assertSame(4, $var2);
     }
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    public function testAllResolvesWithArrayOfResults() {
+        $this->getReactor()->run(function($reactor) {
+            $expected = ['r1' => 42, 'r2' => 41];
+            $actual = (yield 'all' => [
+                'r1' => 42,
+                'r2' => new Success(41),
+            ]);
+            $this->assertSame($expected, $actual);
+        });
+    }
+
+    /**
+     * @expectedException \RuntimeException
+     * @expectedExceptionMessage zanzibar
+     */
+    public function testAllThrowsIfAnyIndividualPromiseFails() {
+        $this->getReactor()->run(function($reactor) {
+            $exception = new \RuntimeException('zanzibar');
+            $promises = [
+                'r1' => new Success(42),
+                'r2' => new Failure($exception),
+                'r3' => new Success(40),
+            ];
+            $results = (yield 'all' => $promises);
+        });
+    }
+
+    public function testSomeReturnsArrayOfErrorsAndResults() {
+        $this->getReactor()->run(function($reactor) {
+            $exception = new \RuntimeException('zanzibar');
+            $promises = [
+                'r1' => new Success(42),
+                'r2' => new Failure($exception),
+                'r3' => new Success(40),
+            ];
+            list($errors, $results) = (yield 'some' => $promises);
+            $this->assertSame(['r2' => $exception], $errors);
+            $this->assertSame(['r1' => 42, 'r3' => 40], $results);
+        });
+    }
+
+    /**
+     * @expectedException \RuntimeException
+     * @expectedExceptionMessage All promises failed
+     */
+    public function testSomeThrowsIfNoPromisesResolveSuccessfully() {
+        $this->getReactor()->run(function($reactor) {
+            $promises = [
+                'r1' => new Failure(new \RuntimeException),
+                'r2' => new Failure(new \RuntimeException),
+            ];
+            list($errors, $results) = (yield 'some' => $promises);
+        });
+    }
+
+    public function testResolvedValueEqualsReturnKeyYield() {
+        $this->getReactor()->run(function($reactor) {
+            $gen = function() {
+                $a = (yield new Success(21));
+                $b = (yield new Success(2));
+                yield 'return' => ($a * $b);
+            };
+
+            $result = (yield 'coroutine' => $gen());
+            $this->assertSame(42, $result);
+        });
+    }
+
+    public function testResolutionFailuresAreThrownIntoGenerator() {
+        $this->getReactor()->run(function($reactor) {
+            $gen = function() {
+                $a = (yield new Success(21));
+                $b = 1;
+                try {
+                    yield new Failure(new \Exception('test'));
+                    $this->fail('Code path should not be reached');
+                } catch (\Exception $e) {
+                    $this->assertSame('test', $e->getMessage());
+                    $b = 2;
+                }
+
+                yield 'return' => ($a * $b);
+            };
+
+            $result = (yield 'coroutine' => $gen());
+            $this->assertSame(42, $result);
+        });
+    }
+
+    /**
+     * @expectedException \Exception
+     * @expectedExceptionMessage When in the chronicle of wasted time
+     */
+    public function testUncaughtGeneratorExceptionFailsResolverPromise() {
+        $this->getReactor()->run(function($reactor) {
+            $gen = function() {
+                yield "pause" => 1;
+                throw new \Exception('When in the chronicle of wasted time');
+                yield "pause" => 1;
+            };
+
+            yield 'coroutine' => $gen();
+        });
+    }
+
+    public function testAllCombinatorResolution() {
+        $this->getReactor()->run(function($reactor) {
+            $gen = function() {
+                list($a, $b) = (yield 'all' => [
+                    new Success(21),
+                    new Success(2),
+                ]);
+                yield 'return' => ($a * $b);
+            };
+
+            $result = (yield 'coroutine' => $gen());
+            $this->assertSame(42, $result);
+        });
+    }
+
+    public function testAllCombinatorResolutionWithNonPromises() {
+        $this->getReactor()->run(function($reactor) {
+            $gen = function() {
+                list($a, $b, $c) = (yield 'all' => [new Success(21), new Success(2), 10]);
+                yield 'return' => ($a * $b * $c);
+            };
+
+            $result = (yield 'coroutine' => $gen());
+            $this->assertSame(420, $result);
+        });
+    }
+
+    /**
+     * @expectedException \Exception
+     * @expectedExceptionMessage When in the chronicle of wasted time
+     */
+    public function testAllCombinatorResolutionThrowsIfAnyOnePromiseFails() {
+        $gen = function() {
+            list($a, $b) = (yield 'all' => [
+                new Success(21),
+                new Failure(new \Exception('When in the chronicle of wasted time')),
+            ]);
+        };
+
+        $this->getReactor()->run(function($reactor) use ($gen) {
+            yield 'coroutine' => $gen();
+        });
+    }
+
+    public function testCombinatorResolvesGeneratorInArray() {
+        $this->getReactor()->run(function($reactor) {
+            $gen1 = function() {
+                yield 'return' => 21;
+            };
+
+            $gen2 = function() use ($gen1) {
+                list($a, $b) = (yield 'all' => [
+                    \Amp\coroutine($gen1(), $reactor),
+                    new Success(2)
+                ]);
+                yield 'return' => ($a * $b);
+            };
+
+            $result = (yield 'coroutine' => $gen2());
+            $this->assertSame(42, $result);
+        });
+    }
+
+    public function testExplicitAllCombinatorResolution() {
+        $this->getReactor()->run(function($reactor) {
+            $gen = function() {
+                list($a, $b, $c) = (yield 'all' => [
+                    new Success(21),
+                    new Success(2),
+                    10
+                ]);
+                yield 'return' => ($a * $b * $c);
+            };
+
+            $result = (yield 'coroutine' => $gen());
+            $this->assertSame(420, $result);
+        });
+    }
+
+    public function testExplicitAnyCombinatorResolution() {
+        $this->getReactor()->run(function($reactor) {
+            $gen = function() {
+                $any = (yield 'any' => [
+                    'a' => new Success(21),
+                    'b' => new Failure(new \Exception('test')),
+                ]);
+                
+                yield 'return' => $any;
+            };
+
+            list($errors, $results) = (yield 'coroutine' => $gen());
+            $this->assertSame('test', $errors['b']->getMessage());
+            $this->assertSame(21, $results['a']);
+        });
+    }
+
+    /**
+     * @expectedException \RuntimeException
+     * @expectedExceptionMessage All promises failed
+     */
+    public function testExplicitSomeCombinatorResolutionFailsOnError() {
+        $this->getReactor()->run(function($reactor) {
+            $gen = function() {
+                yield 'some' => [
+                    'r1' => new Failure(new \RuntimeException),
+                    'r2' => new Failure(new \RuntimeException),
+                ];
+            };
+            yield 'coroutine' => $gen();
+        });
+    }
+
+    /**
+     * @expectedException \DomainException
+     * @expectedExceptionMessage "some" yield command expects array; string yielded
+     */
+    public function testExplicitCombinatorResolutionFailsIfNonArrayYielded() {
+        $this->getReactor()->run(function($reactor) {
+            $gen = function() {
+                yield 'some' => 'test';
+            };
+            yield 'coroutine' => $gen();
+        });
+    }
+
+    public function testCallableBindYield() {
+        $this->getReactor()->run(function($reactor) {
+            // Register a repeating callback so the reactor run loop doesn't break
+            // without our intervention.
+            $repeatWatcherId = (yield 'repeat' => [function(){}, 1000]);
+
+            $func = function() use ($repeatWatcherId) {
+                yield "cancel" => $repeatWatcherId;
+            };
+
+            $boundFunc = (yield "bind" => $func);
+
+            // Because this Generator function is bound to the reactor it should be
+            // automatically resolved and our repeating watcher should be cancelled
+            // allowing the reactor to stop running.
+            $result = $boundFunc();
+            $this->assertInstanceOf('Amp\\Promise', $result);
+        });
+    }
+    
+    public function testExplicitImmediatelyYieldResolution() {
+        $this->getReactor()->run(function($reactor) {
+            $gen = function() {
+                $var = null;
+                yield 'immediately' => function() use (&$var) { $var = 42; };
+                yield 'pause' => 100; // pause for 100ms so the immediately callback executes
+                yield 'return' => $var;
+            };
+            $result = (yield 'coroutine' => $gen());
+            $this->assertSame(42, $result);
+        });
+    }
+
+    public function testExplicitOnceYieldResolution() {
+        $this->getReactor()->run(function($reactor) {
+            $gen = function() {
+                $var = null;
+                yield 'once' => [function() use (&$var) { $var = 42; }, $msDelay = 1];
+                yield 'pause' => 100; // pause for 100ms so the once callback executes
+                yield 'return' => $var;
+            };
+            $result = (yield 'coroutine' => $gen());
+            $this->assertSame(42, $result);
+        });
+    }
+
+    public function testExplicitRepeatYieldResolution() {
+        $this->getReactor()->run(function($reactor) {
+            $var = null;
+            $repeatFunc = function($reactor, $watcherId) use (&$var) {
+                $var = 1;
+                yield 'cancel' => $watcherId;
+                $var++;
+            };
+
+            $gen = function() use (&$var, $repeatFunc) {
+                yield 'repeat' => [$repeatFunc, $msDelay = 1];
+                yield 'pause'   => 100; // pause for 100ms so we can be sure the repeat callback executes
+                yield 'return' => $var;
+            };
+
+            $result = (yield 'coroutine' => $gen());
+            $this->assertSame(2, $result);
+        });
+    }
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
 }